+++ examples/ReSpeaker 2mic HAT/DOCS.md
@@ -150,7 +150,7 @@ If it is not `1000`, update the `user:` field in `compose.yml` to match.
 ### Step 4 — File structure
 
 ```
-respeaker_2mic_hat/
+ReSpeaker 2mic HAT/
 ├── Dockerfile
 ├── compose.yml
 ├── requirements.txt
